Split transaction into multiple files.
Renamed a few type names as well.
